fold ServerSelector::Selectable into ServerSelector::Base
