+++ lib/mongo/server_selector/primary.rb
@@ -21,7 +21,6 @@ module Mongo
     #
     # @since 2.0.0
     class Primary < Base
-      include Selectable
 
       # Name of the this read preference in the server's format.
       #
