+++ lib/mongo/server_selector/secondary.rb
@@ -21,7 +21,6 @@ module Mongo
     #
     # @since 2.0.0
     class Secondary < Base
-      include Selectable
 
       # Name of the this read preference in the server's format.
       #
